# src/swagger/swagger.json
{
	"components": {
...
		"schemas": {
...
			"TSOAField": {
				"enum": [
					"TITLE",
					"TYPE",
					"VARIANT_COMPARE_AT_PRICE",
					"VARIANT_PRICE",
					"VENDOR"
				],
				"type": "string"
			},
			"TSOAConditionEnum": {
				"enum": [
					"CONTAINS",
					"ENDS_WITH",
					"EQUALS",
					"GREATER_THAN",
					"LESS_THAN",
					"STARTS_WITH"
				],
				"type": "string"
			},
			"TSOAConditionElement": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"field": {
						"$ref": "#/components/schemas/TSOAField"
					},
					"condition": {
						"$ref": "#/components/schemas/TSOAConditionEnum"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"field",
					"condition",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TSOACollection": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"mustMatchAll": {
						"type": "boolean"
					},
					"date": {
						"type": "string"
					},
					"deleted": {
						"type": "boolean"
					},
					"chartImage": {
						"type": "string",
						"nullable": true
					},
					"conditions": {
						"items": {
							"$ref": "#/components/schemas/TSOAConditionElement"
						},
						"type": "array"
					},
					"createdBy": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"__v": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "string"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"title",
					"description",
					"status",
					"mustMatchAll",
					"date",
					"deleted",
					"conditions",
					"createdBy",
					"createdAt",
					"updatedAt",
					"slug",
					"__v"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TSOASubCategory": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"deleted": {
						"type": "boolean"
					},
					"name": {
						"type": "string"
					},
					"level": {
						"type": "number",
						"format": "double"
					},
					"createdBy": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"__v": {
						"type": "number",
						"format": "double"
					},
					"date": {
						"type": "string"
					},
					"collections": {
						"items": {
							"$ref": "#/components/schemas/TSOACollection"
						},
						"type": "array"
					}
				},
				"required": [
					"_id",
					"status",
					"deleted",
					"name",
					"level",
					"createdBy",
					"createdAt",
					"updatedAt",
					"slug",
					"__v",
					"date"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TSOACategory": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"subCategories": {
						"items": {
							"$ref": "#/components/schemas/TSOASubCategory"
						},
						"type": "array"
					},
					"status": {
						"type": "boolean"
					},
					"deleted": {
						"type": "boolean"
					},
					"name": {
						"type": "string"
					},
					"level": {
						"type": "number",
						"format": "double"
					},
					"createdBy": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"__v": {
						"type": "number",
						"format": "double"
					},
					"date": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"subCategories",
					"status",
					"deleted",
					"name",
					"level",
					"createdBy",
					"createdAt",
					"updatedAt",
					"slug",
					"__v",
					"date"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoriesResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"count": {
								"type": "number",
								"format": "double"
							},
							"docs": {
								"items": {
									"$ref": "#/components/schemas/TSOACategory"
								},
								"type": "array"
							}
						},
						"required": [
							"count",
							"docs"
						],
						"type": "object"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "boolean"
					}
				},
				"required": [
					"message",
					"data",
					"statusCode",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoryResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/TSOACategory"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "boolean"
					}
				},
				"required": [
					"message",
					"data",
					"statusCode",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CollectionResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/TSOACollection"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "boolean"
					}
				},
				"required": [
					"message",
					"data",
					"statusCode",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CollectionReqBody": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"mustMatchAll": {
						"type": "boolean"
					},
					"chartImage": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"conditions": {
						"items": {
							"properties": {
								"value": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"type": "number",
											"format": "double"
										}
									]
								},
								"condition": {
									"type": "string"
								},
								"field": {
									"type": "string"
								}
							},
							"required": [
								"value",
								"condition",
								"field"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"description",
					"type",
					"mustMatchAll",
					"chartImage",
					"image",
					"conditions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CollectionsResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"count": {
								"type": "number",
								"format": "double"
							},
							"docs": {
								"items": {
									"$ref": "#/components/schemas/TSOACollection"
								},
								"type": "array"
							}
						},
						"required": [
							"count",
							"docs"
						],
						"type": "object"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "boolean"
					}
				},
				"required": [
					"message",
					"data",
					"statusCode",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "Add Bearer prefix on jwt token"
			}
		}
	},
	"info": {
		"title": "Market Place Apis Documentation",
		"version": "1.0.0",
		"description": "Market Place Node Backend Apis and Documentation",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"email": "marketplace@gmail.com",
			"name": "marketplace",
			"url": "http://marketplace.in"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/categories": {
			"get": {
				"operationId": "GetMainCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoriesResponse"
								}
							}
						}
					}
				},
				"summary": "fetch categories",
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "createdAt",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateMainCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryResponse"
								}
							}
						}
					}
				},
				"summary": "Create new category",
				"tags": [
					"Category"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"image": {
										"type": "string"
									},
									"status": {
										"type": "boolean"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"image",
									"status",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/categories/{categoryId}": {
			"get": {
				"operationId": "GetMainCategoryById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryResponse"
								}
							}
						}
					}
				},
				"summary": "fetch single category details",
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateMainCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryResponse"
								}
							}
						}
					}
				},
				"summary": "Update existing category",
				"tags": [
					"Category"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"image": {
										"type": "string"
									},
									"status": {
										"type": "boolean"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"image",
									"status",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateMainCategoryStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryResponse"
								}
							}
						}
					}
				},
				"summary": "Update existing category active status",
				"tags": [
					"Category"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"status": {
										"type": "boolean"
									}
								},
								"required": [
									"status"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMainCategoryStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryResponse"
								}
							}
						}
					}
				},
				"summary": "Delete existing category",
				"tags": [
					"Category"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/categories/by-admin": {
			"get": {
				"operationId": "GetMainCategoriesByAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoriesResponse"
								}
							}
						}
					}
				},
				"summary": "fetch categories by admin",
				"tags": [
					"Category"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "createdAt",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					}
				]
			}
		},
...
		"/collections": {
			"post": {
				"operationId": "CreateCollection",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CollectionResponse"
								}
							}
						}
					}
				},
				"summary": "Create new collection",
				"tags": [
					"Collections"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CollectionReqBody"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetCollections",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CollectionsResponse"
								}
							}
						}
					}
				},
				"summary": "fetch collections",
				"tags": [
					"Collections"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "createdAt",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "isFeatured",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/collections/{collectionId}": {
			"put": {
				"operationId": "UpdateCollection",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CollectionResponse"
								}
							}
						}
					}
				},
				"summary": "Update existing collection",
				"tags": [
					"Collections"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "collectionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CollectionReqBody"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateCollectionStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CollectionResponse"
								}
							}
						}
					}
				},
				"summary": "Update existing collection active status",
				"tags": [
					"Collections"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "collectionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"status": {
										"type": "boolean"
									}
								},
								"required": [
									"status"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCollection",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CollectionResponse"
								}
							}
						}
					}
				},
				"summary": "Delete existing collection",
				"tags": [
					"Collections"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "collectionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "GetCollectionById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CollectionResponse"
								}
							}
						}
					}
				},
				"summary": "fetch single collection details",
				"tags": [
					"Collections"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "collectionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/collections/featured/{collectionId}": {
			"patch": {
				"operationId": "UpdateCollectionFeaturedStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CollectionResponse"
								}
							}
						}
					}
				},
				"summary": "Update existing collection featured status",
				"tags": [
					"Collections"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "collectionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"isFeatured": {
										"type": "boolean"
									}
								},
								"required": [
									"isFeatured"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/collections/by-admin": {
			"get": {
				"operationId": "GetCollectionsByAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CollectionsResponse"
								}
							}
						}
					}
				},
				"summary": "fetch collections by admin",
				"tags": [
					"Collections"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "createdAt",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "isFeatured",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}